+++ cmd/sm-config-edit/frontend/src/components/ActionsEditor.svelte
@@ -22,6 +22,7 @@
       groups: [],
       noWait: false,
       interactive: true,
+      requiresPin: false,
     } as unknown as configedit.ActionDTO
   }